Added option to add unmatched shows automatically.    This will search TVDB for the show. On finding it, if a show with the same TVDB ID already exists it will add an alternate name. Else it will add a new show and update it. Added option to unlock a show when a match is made.    This will match the show as normal. If it is locked, then it will unlock the show and update its episode information. Added option to lock shows that haven't aired in 3 weeks.    After an update, this will lock any show that hasn't had an episode in the last 3 weeks. When it starts airing again, the option above will unlock it.
The purpose of those 2 features is for automatic management of the active shows that need updating, while not bothering to update shows that are airing new episodes.

Now shows the current verison on the UI header.
Add events from the XML class that fire when show's or settings are changed so the UI can update accordingly.

Implemented the RenameIfExists setting. This searches the desintation directory for the file, it is already exists under a different name, it will rename it instead of moving it, unless the new file contains an overwrite keyword.